+++ zwave_me_server/run.sh
@@ -7,6 +7,11 @@ device=/dev/ttyACM0
 # Do not use this function if more than Z-Wave 1 interface is used
 forceDevice=false
 
+# By default remote access is on
+remote_access=true
+# And tech support access is not
+remote_support_access=false
+
 # Default config path
 defCJ=/opt/z-way-server/automation/defaultConfigs/config.json
 
@@ -14,7 +19,10 @@ defCJ=/opt/z-way-server/automation/defaultConfigs/config.json
 if [ -f "/data/options.json" ]; then
   # Get device path from Configuration tab of addon
   device=$(grep -Eo '/dev/tty[A-Z]*[0-9]' /data/options.json)
-  forceDevice=$(grep "configjson_device_replace" /data/options.json | cut -d ":" -f 2)
+  forceDevice=$(grep "configjson_device_replace" /data/options.json | cut -d ":" -f 2 | tr -d ',')
+  remote_access=$(grep "remote_access" /data/options.json | cut -d ":" -f 2 | tr -d ',')
+  remote_support_access=$(grep "remote_support_access" /data/options.json | cut -d ":" -f 2 | tr -d ',')
+  zbw_password=$(grep "zbw_password" /data/options.json | cut -d ":" -f 2 | tr -d '," ')
 fi
 
 # Change device path in /defailtConfig/config.json
@@ -61,11 +69,32 @@ ln -sf /data/opt/z-way-server/configs/ /opt/z-way-server/
 
 # Use homeassistant.local in ZBW (for find.z-wave.me) instead of the local IP as we can't get the IP outside the docker
 echo homeassistant.local > /etc/zbw/local_ips
+echo "8083" > /etc/zbw/local_port
+
+# Check if remote access for support enable
+if [ $remote_support_access ]; then
+  touch /etc/zbw/flags/forward_ssh
+else
+  rm /etc/zbw/flags/forward_ssh
+fi
+
+# Change support user default password
+if [ -n "$zbw_password" ]; then
+  echo "support:$zbw_password" | chpasswd
+fi
+
+# If remote access enable start zbw_connect
+if [ $remote_access ]; then
+  rm /etc/zbw/flags/no_connection
+else
+  touch /etc/zbw/flags/no_connection
+fi
 
 # Starting Z-Way services
 /etc/init.d/mongoose start
 /etc/init.d/z-way-server start
 /etc/init.d/zbw_connect start
+service ssh start
 
 # colored log output
 tail --pid `cat /var/run/z-way-server.pid` -F /var/log/z-way-server.log | /opt/z-way-server/colorize-log.sh