+++ zwave_me_server/CHANGELOG.md
@@ -1,4 +1,12 @@
 <!-- https://developers.home-assistant.io/docs/add-ons/presentation#keeping-a-changelog -->
+## 1.1.13
+
+- Updated the z-way version to 4.0.3 
+- Add checkbox for remote access
+- Add tech support access checkbox and password (login support, default password razberry)
+- Fix not showing homeassistant.local in find.z-wave.me
+- Opened port 8883 for MQTT
+
 
 ## 1.1.12